requirements.txt
fastapi
uvicorn
python-telegram-bot
pandas
codaio
pydantic
python-dotenv
setuptools
pyppeteer
uvicorn[standard]
openpyxl
httpx
selenium
pytest-asyncio
psycopg2-binary